完成了桌面小部件的功能，可以选择任意菜谱显示

+++ app/src/main/java/com/example/android/bakingapp/Model/IngredientModel.java
@@ -31,7 +31,7 @@ public class IngredientModel {
      * @param measure  单位
      * @param ingredients  用料
      */
-    public IngredientModel(int quantity,String measure,String ingredients) {
+    public IngredientModel(float quantity,String measure,String ingredients) {
         this.mIngredientMeasure = measure;
         this.mIngredientQuantity = quantity;
         this.mIngredients = ingredients;
